Additional conditions,  reports, patient groups
Added additional health conditions/symptoms to select
Added additional stat reports:
-Percentage breakdown of patient ethnicities
-Health outcomes/conditions
-Percentage breakdown of patient groups
Doctors and nurses can now specify a "group" that a patient belongs to
(General, Intensive Care, Surgery, Recovery, Cancer Treatment).  This
can be specified when viewing the Health Conditions of a Patient as a
Doctor or Nurse.  By default, a Patient is placed into the "General"
group.